Made several minor changes.
Made several minor changes as well as advancing in the idea of what the blog should be. I am still figuring out multiple aspects in what the blog will feature and how it can be presentable while still being on the balance of what I want it to be as well as what it should be.